requirements.txt
streamlit==1.39.0
pandas==2.2.2
openpyxl==3.1.5
pdfplumber==0.11.4
pdf2image==1.17.0
pillow==10.4.0
opencv-python==4.6.0.66
paddleocr==2.7.0.3
paddlepaddle==2.6.2
numpy==1.26.4
azure-ai-formrecognizer==3.3.3
python-dotenv==1.0.1
